+++ scout/adapter/mongo/disease_terms.py
@@ -15,11 +15,11 @@ DISEASE_FILTER_PROJECT = {"hpo_terms": 0, "genes": 0}
 class DiagnosisHandler(object):
     """Class for handling OMIM and disease-related database objects"""
 
-    def query_omim(self, query: str = None, limit: int = None) -> Iterable:
-        """Return all OMIM terms
+    def query_disease(self, query: str = None, source: str = "OMIM", limit: int = None) -> Iterable:
+        """Return all disease_terms
 
-        If a query is sent omim_id will try to match with regex on term or
-        description.
+        If a query is sent it will try to match with regex on term or
+        description. If source is sent it will be used to limit the results
         """
 
         query_dict = {}
@@ -30,8 +30,19 @@ class DiagnosisHandler(object):
                     {"description": {"$regex": query, "$options": "i"}},
                 ]
             }
+            # If source is specified, add this restriction to the query
+            if source:
+                query_dict = {
+                    "$and": [
+                        query_dict,
+                        {"source": source},
+                    ]
+                }
+        elif source:
+            query_dict = {"source": source}
 
         limit = limit or int(10e10)
+
         res = (
             self.disease_term_collection.find(query_dict, DISEASE_FILTER_PROJECT)
             .limit(limit)
@@ -73,7 +84,7 @@ class DiagnosisHandler(object):
             "disease_id", ASCENDING
         )
 
-    def omim_to_genes(self, omim_obj: dict) -> List[dict]:
+    def disease_to_genes(self, omim_obj: dict) -> List[dict]:
         """Gets all genes associated to an OMIM term."""
         gene_objs = []
         if omim_obj:
@@ -93,25 +104,44 @@ class DiagnosisHandler(object):
 
     def disease_terms(
         self,
-        hgnc_id: Optional[int] = None,
+        source: Optional[str] = "OMIM",
         filter_project: Optional[dict] = DISEASE_FILTER_PROJECT,
     ) -> list:
-        """Return all disease terms for a gene HGNC ID. Optionally filter the returned key/values using filter_project.
-        By default, do not return disease-associated genes and HPO terms."""
+        """Return all disease terms optionally from only one source and filtered the returned key/values using filter_project. By default do not return disease-associated genes and HPO terms."""
         query = {}
-        if hgnc_id:
-            LOG.debug("Fetching all diseases for gene %s", hgnc_id)
-            query = {"genes": hgnc_id}
+        if source:
+            LOG.debug(f"Fetching all {source} diseases")
+            query = {"source": source}
         else:
             LOG.info("Fetching all disease terms")
 
         return list(self.disease_term_collection.find(query, filter_project))
+
+    def disease_terms_by_gene(
+        self,
+        hgnc_id: [int],
+        source: Optional[str] = None,
+        filter_project: Optional[dict] = DISEASE_FILTER_PROJECT,
+    ) -> list:
+        """Return all disease terms for a gene HGNC ID. Optionally filter the returned key/values using filter_project.
+        By default, do not return disease-associated genes and HPO terms."""
+        if source:
+            LOG.debug(f"Fetching all {source} diseases for gene { hgnc_id}")
+            query = {"$and": [{"source": source}, {"genes": hgnc_id}]}
+        else:
+            LOG.debug(f"Fetching all diseases for gene { hgnc_id}")
+            query = {"genes": hgnc_id}
+
+        return list(self.disease_term_collection.find(query, filter_project))
+
+    def disease_terminology_count(self) -> list:
+        """Return the count of disease_terms for each source in the db"""
+        query = {"$group": {"_id": "$source", "count": {"$sum": 1}}}
+        return list(self.disease_term_collection.aggregate([query]))
 
     def load_disease_term(self, disease_obj: dict):
         """Load a disease term into the database-"""
         try:
             self.disease_term_collection.insert_one(disease_obj)
         except DuplicateKeyError:
-            raise IntegrityError(
-                "Disease term %s already exists in database".format(disease_obj["_id"])
-            )
+            raise IntegrityError(f"Disease term {disease_obj['_id']} already exists in database")