+++ scout/server/blueprints/cases/views.py
@@ -631,7 +631,7 @@ def omimterms():
     query = request.args.get("query")
     if query is None:
         return abort(500)
-    terms = store.query_omim(query=query)
+    terms = store.query_disease(query=query)
     json_terms = [
         {"name": "{} | {}".format(term["_id"], term["description"]), "id": term["_id"]}
         for term in terms[:7]