+++ scout/server/blueprints/diagnoses/controllers.py
@@ -7,19 +7,19 @@ from scout.server.links import disease_link
 LOG = logging.getLogger(__name__)
 
 
-def disease_entry(store, disease_id):
-    """Retrieve specific info for an OMIM term at the gene level
+def disease_entry(store, disease_id) -> dict:
+    """Retrieve specific info for a disease term
 
     Args:
         store(obj): an adapter to the scout database
         disease_id(str): a disease_id
 
     Returns:
-        omim_obj(obj): an OMIM term containing description and genes
+        disease_obj(obj): a disease term containing description and genes
     """
 
     disease_obj = store.disease_term(disease_identifier=disease_id, filter_project={})
-    disease_obj["genes_complete"] = store.omim_to_genes(disease_obj)
+    disease_obj["genes_complete"] = store.disease_to_genes(disease_obj)
     disease_obj["hpo_complete"] = [
         store.hpo_term(hpo_id) for hpo_id in disease_obj.get("hpo_terms", [])
     ]